Added a searchbar functionality to the Calendar.
Added Remaining time to Upcomming events at the Calendar.
Added scroll funtionality to the tab bar for a better responsivity on smaller devices.
Added a small panel to profiles with a picture, name and profile photo.
Fixed a lot of textual and visual bugs with the Profiles.

Added a date and a local to Tournament for better management, both on front end and back end.

Added a document with a draft of the Injury model.

Added CanvasJs to the project. CanvasJs is used to create graphs with Javascript and Angular. (unused right now)